individual folders for each component for scoped style sheets
