+++ app/src/components/Registration/TextArea/TextArea.tsx
